add shared truth table feature module
